DEprot.RMSE class

DEprot.RMSE class

Slots

original.DEprot.object

Object of class DEprot used to compute the RMSE. Class: "ANY".

percentage.test

Percentage of the total proteins that should be used to perform the comparisons. Class: "ANY".

seed

Seed used for the randomization. Class: "ANY".

fraction.missing.values

Fraction of missing values in the original table. Class: "ANY".

test.dataset

Subset of the original table used for the comparisons. Class: "ANY".

imputed.objects

List of the output of impute.counts (class DEprot) using the different imputation methods. Class: "ANY".

RMSE.tables

List of data.frames, one per tested imputation method, containing the coordinates of the imputed values and including the following columns (Class: "ANY".):

row.id: id of the row (protein)

col.id: id of the column (sample)

expected.values: value measured in the experiment

imputation.method: id of the method used for the imputation

imputed.values: value imputed by DEprot

residuals: difference of the the values, imputed - expected

sq.residuals: the squared value of the residuals

contrasts

List of contrasts. each contrast is a vector indicating, in the order: metadata.table.column - groupA - groupB; (groupA / group B). Class: "ANY".

RMSE.scores

A table indicating the methods and the corresponding RMSE score. Class: "ANY".

correlation.plots

A list of ggplot objects with the correlation between observed/expected and imputed values. Class: "ANY".

density.residuals

A list of ggplot objects depicting the distribution of the residuals. Class: "ANY".