DEprot.analyses class

DEprot.analyses class

Slots

metadata

The data.frame corresponding to the metadata table describing the samples. Class: "ANY".

raw.counts

Numeric matrix (rows: proteins, columns: samples) of the raw counts. Class: "ANY".

norm.counts

Numeric matrix (rows: proteins, columns: samples) of the normalized counts. Class: "ANY".

imputed.counts

Numeric matrix (rows: proteins, columns: samples) of the imputed counts. Class: "ANY".

log.base

Numeric value indicating the base of the logarithm expressing the values in the loaded data. Class: "ANY".

log.transformed

Logical value indicating whether the data are log-transformed or not. Class: "logical".

imputed

Logical value indicating whether the data are imputed. Class: "logical".

imputation

String (or any other class) value indicating the imputation method. Class: "ANY".

normalized

Logical value indicating whether the data are normalized. Class: "logical".

normalization.method

String (or any other class) value indicating the normalization method. Class: "ANY". Class: "ANY".

boxplot.raw

Ggplot object showing the distribution of the raw values per sample. Class: "ANY".

boxplot.norm

Ggplot object showing the distribution of the normalized values per sample. Class: "ANY".

boxplot.imputed

Ggplot object showing the distribution of the imputed values per sample. Class: "ANY".

analyses.result.list

List containing the differential results for each contrast. Class: "ANY". The list contains the following elements:

results: a data.frame containing the results of the analyses; includes average expression of each group, basemean, foldchange, pvalue and p.adj, differential.status

n.diff: a summary table showing the number of proteins in each differential expression status (up/down/unresponsive, null)

PCA.data: output of perform.PCA for the subset of samples analyzed in a specific contrast

PCA.plots: combination of 3 plots: scatter PC1-vs-PC2, scatter PC2-vs-PC3, and cumulative bar plot

correlations: combination of Pearson and Spearman correlation heatmaps (obtained by plot.correlation.heatmap) for the subset of samples analyzed in a specific contrast

volcano: volcano plot showing the log2(FoldChange) x -log10(p.adjusted) of differential expression results; it can be regenerated using plot.volcano

MA.plot: MA-plot showing the log2(basemean) x log2(FoldChange) of differential expression results; it can be regenerated using plot.MA

List

of contrasts. each contrast is a vector indicating, in the order: metadata.table.column - groupA - groupB; (groupA / group B). Class: "ANY".

differential.analyses.params

List of parameters used to run the differential analyses (fold change thresholds, p-value threshold, p-adjustement method, etc.). Class: "ANY".