DEprot.correlation class

DEprot.correlation class

Slots

heatmap:

ggplot object corresponding to the correlation heatmap. Class: "ANY".

corr.metadata:

metadata of the samples used in the correlation (subset of the original DEprot@metadata. Class: "ANY".

sample.subset:

vector containing the list of samples analyzed. Class: "ANY".

data.used:

vector indicating the type of counts used (imputed, normalized, raw). Class: "ANY".

corr.matrix:

the correlation matrix on which the heatmap is base on. Class: "ANY".

distance:

object of class dist corresponding to the output of as.dist(1 - correlation.matrix). Class: "ANY".

cluster:

hclust object generated by hclust(d = as.dist(1 - correlation.matrix), method = clustering.method). Class: "ANY".