NES.plot

GSEA analyses NES plotter. Plotting of a bar plot indicating the Normalized Enrichment Score (NES) of a gsea result. Bars will have different colors by positive or negative enrichment and transparency proportional to their p-value.

NES.plot(
  enrichResult,
  pos.NES.label = "+NES",
  neg.NES.label = "-NES",
  pos.NES.color = "steelblue",
  neg.NES.color = "orange",
  string.pattern.to.remove = "HALLMARK|GOBP",
  alpha.range = c(0.3, 1),
  add.counts = TRUE,
  perc.bleeding.x = 8,
  axes.text.size = 10,
  title = "NES enrichments"
)

Arguments

enrichResult

A DEprot.enrichResult object, as generated by geneset.enrichment DEprot function, or a gseaResult object as generate by clusterprofiler.

pos.NES.label

String for the label of sets enriched in the group with positive ranking feature scores. Default: "+NES".

neg.NES.label

String for the label of sets enriched in the group with negative ranking feature scores. Default: "-NES".

pos.NES.color

String for the color of sets enriched in the group with positive ranking feature scores. Default: "steelblue".

neg.NES.color

String for the color of sets enriched in the group with negative ranking feature scores. Default: "orange".

string.pattern.to.remove

String with a regular expression of a pattern to be removed from the set names. Default: "HALLMARK|GOBP".

alpha.range

Numeric vector of length 2 indicating minimum and maximum value for the transparency. Individual values must be a number between 0 and 1. Default: c(0.3, 1).

add.counts

Logic value to indicate whether add labels with the set size counts. Default: "TRUE".

perc.bleeding.x

Numeric value indicating the percentage of the full x.axis range to add on the left on the right. Useful when labels are falling outside for the x-max. Default: 8 (%).

axes.text.size

Numeric value indicating the font size of the axis text. Default: 10.

title

String indicating the title of the plot. Default: "NES enrichments".

Value

A ggplot object.

Author

Sebastian Gregoricchio