.check.aes.columns

Internal. Validates the columns requested for the color/shape/label aesthetics of the ordination plots and prepares the corresponding vectors and legend labels.

.check.aes.columns(
  coordinates,
  color.column,
  shape.column = NULL,
  label.column = NULL,
  axis.prefix = "PC"
)

Arguments

coordinates

data.frame combining the ordination coordinates and the metadata (i.e. the PCs slot of a DEprot.PCA or the PCos slot of a DEprot.PCoA).

color.column

String indicating the name of the column to use as factor for the dot colors.

shape.column

String indicating the name of the column to use as factor for the dot shapes, or NULL.

label.column

String indicating the name of the column to use as label of the dots, or NULL.

axis.prefix

String indicating the prefix of the coordinate columns ('PC' or 'PCo').

Value

A list with the elements: shape.scores, shape.label, label.column, show.labels.