get.results

Simplifies the access to the differential expression results table. When the object contains a protein annotation table (protein.info slot, see load.counts2 and add.protein.info), the annotation columns can be appended to the results.

get.results(
  DEprot.analyses.object,
  contrast = 1,
  protein.info.columns = "none",
  protein.info.prefix = NULL
)

Arguments

DEprot.analyses.object

An object of class DEprot.analyses.

contrast

Number indicating the position of the contrast to use for the plotting.

protein.info.columns

String vector indicating which columns of the protein.info slot should be appended to the results table. Two keywords are available: "none", to return the results alone, and "all", to append the whole annotation. Any other value is interpreted as the names of the columns to append. Ignored when no annotation is stored in the object. Default: "none".

protein.info.prefix

String to prepend to the names of the appended annotation columns (e.g., "info."), useful to distinguish them from the columns of the results. If NULL (default), the original names are kept and only the ones colliding with the results columns are made unique.

Value

A data frame.

See also

Author

Sebastian Gregoricchio

Examples

# Results alone
get.results(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
            contrast = 1)
#>       prot.id basemean.log2 log2.mean.FBS log2.mean.6h.DMSO
#> 1   protein.1      18.54165      18.26954          18.81376
#> 2   protein.2      22.76659      22.43601          23.09716
#> 3   protein.3      16.20940      16.57051          15.84830
#> 4   protein.4      21.19592      21.16429          21.22754
#> 5   protein.5      19.30638      19.18238          19.43038
#> 6   protein.6      15.49009      15.63217          15.34802
#> 7   protein.7      17.00778      17.02135          16.99422
#> 8   protein.8      12.11628      12.69995          11.53260
#> 9   protein.9      11.59300      12.08864          11.09737
#> 10 protein.10      12.84557      13.11600          12.57513
#> 11 protein.11      14.22426      14.25200          14.19653
#> 12 protein.12      14.18435      14.40897          13.95974
#> 13 protein.14      18.17955      18.29240          18.06671
#> 14 protein.15      15.15596      15.29837          15.01355
#> 15 protein.16      16.30529      16.39731          16.21326
#> 16 protein.17      13.42623      14.22905          12.62341
#> 17 protein.18      17.70510      17.48411          17.92609
#> 18 protein.19      15.97148      15.95095          15.99202
#> 19 protein.20      19.39488      19.48232          19.30743
#> 20 protein.21      18.47346      18.47229          18.47463
#> 21 protein.22      16.49923      16.49822          16.50023
#> 22 protein.23      15.42045      15.55763          15.28327
#> 23 protein.24      14.96077      15.07249          14.84904
#> 24 protein.25      18.93194      18.66327          19.20061
#> 25 protein.26      17.13326      16.50530          17.76121
#> 26 protein.27      17.49123      17.68725          17.29522
#> 27 protein.28      11.30874      11.15241          11.46507
#> 28 protein.29      16.36207      16.73584          15.98829
#> 29 protein.30      17.36296      17.05715          17.66877
#> 30 protein.31      12.96287      13.30513          12.62060
#> 31 protein.32      13.33568      13.13449          13.53687
#> 32 protein.33      15.43228      15.42325          15.44131
#> 33 protein.34      12.63760      11.88392          13.39128
#> 34 protein.35      14.31802      14.63020          14.00584
#> 35 protein.36      16.00716      16.04804          15.96629
#> 36 protein.37      19.58236      19.50252          19.66220
#> 37 protein.38      16.53489      16.62487          16.44492
#> 38 protein.39      17.04712      17.27062          16.82361
#> 39 protein.40      16.23806      16.25504          16.22108
#> 40 protein.41      16.32023      16.18395          16.45652
#> 41 protein.42      15.59224      15.31647          15.86801
#> 42 protein.43      16.78640      16.57529          16.99751
#> 43 protein.44      14.07439      13.36062          14.78817
#> 44 protein.45      15.83634      16.10593          15.56675
#> 45 protein.46      15.08760      14.95140          15.22380
#> 46 protein.47      11.84487      12.15713          11.53260
#> 47 protein.48      17.49097      18.35534          16.62660
#> 48 protein.49      17.19600      17.46597          16.92602
#> 49 protein.50      15.79326      15.90305          15.68347
#> 50 protein.51      15.32080      15.76659          14.87500
#>    log2.Fold_FBS.vs.6h.DMSO      p.value        padj   statistic       df
#> 1              -0.544224676 2.788667e-02 0.094452391 -2.50194806 11.95118
#> 2              -0.661148700 3.908730e-04 0.004885913 -4.86794370 11.95118
#> 3               0.722211468 2.579070e-03 0.018421927  3.79324426 11.95118
#> 4              -0.063254105 5.916181e-01 0.704307234 -0.55125200 11.95118
#> 5              -0.247994864 1.229210e-01 0.219501792 -1.65991770 11.95118
#> 6               0.284145196 1.299883e-01 0.224117783  1.62610346 11.95118
#> 7               0.027126811 9.120523e-01 0.954600916  0.11281216 11.95118
#> 8               1.167347745 5.376507e-02 0.128012081  2.13912514 11.95118
#> 9               0.991273547 3.709303e-04 0.004885913  4.89900805 11.95118
#> 10              0.540865019 5.051674e-02 0.126291840  2.17405019 11.95118
#> 11              0.055474587 8.928236e-01 0.954600916  0.13763107 11.95118
#> 12              0.449233938 1.781768e-01 0.262024729  1.43056764 11.95118
#> 13              0.225695196 8.696017e-02 0.183348825  1.86468854 11.95118
#> 14              0.284825720 1.729014e-01 0.261971825  1.44959543 11.95118
#> 15              0.184057322 3.450198e-01 0.442333030  0.98314791 11.95118
#> 16              1.605633047 8.486982e-05 0.002121745  5.80929006 11.95118
#> 17             -0.441985756 4.686072e-02 0.123317671 -2.21599204 11.95118
#> 18             -0.041072472 7.912847e-01 0.899187125 -0.27064076 11.95118
#> 19              0.174882092 1.135883e-01 0.218438960  1.70730119 11.95118
#> 20             -0.002342062 9.879577e-01 0.988477955 -0.01541206 11.95118
#> 21             -0.002006187 9.884780e-01 0.988477955 -0.01474622 11.95118
#> 22              0.274353614 1.621873e-01 0.261592441  1.48980115 11.95118
#> 23              0.223451682 4.352896e-01 0.530840975  0.80725895 11.95118
#> 24             -0.537343331 3.418833e-02 0.100553906 -2.39038682 11.95118
#> 25             -1.255916962 1.991072e-03 0.018421927 -3.93631105 11.95118
#> 26              0.392025553 1.264107e-02 0.052671111  2.93046720 11.95118
#> 27             -0.312661079 2.299120e-01 0.310691916 -1.26519923 11.95118
#> 28              0.747543162 9.025440e-03 0.050141332  3.11210100 11.95118
#> 29             -0.611619251 3.578412e-03 0.022365077 -3.61360374 11.95118
#> 30              0.684529841 8.800744e-02 0.183348825  1.85772225 11.95118
#> 31             -0.402377887 1.581353e-01 0.261592441 -1.50559713 11.95118
#> 32             -0.018062115 9.164169e-01 0.954600916 -0.10718969 11.95118
#> 33             -1.507356091 1.086806e-02 0.050353715 -3.01193758 11.95118
#> 34              0.624360902 4.356655e-02 0.121018205  2.25653070 11.95118
#> 35              0.081753522 7.897580e-01 0.899187125  0.27267397 11.95118
#> 36             -0.159672887 2.159763e-01 0.300161028 -1.30647011 11.95118
#> 37              0.179953720 3.379315e-01 0.442333030  0.99827111 11.95118
#> 38              0.447016255 2.306848e-02 0.088724919  2.60522179 11.95118
#> 39              0.033956247 8.369197e-01 0.929910803  0.21037408 11.95118
#> 40             -0.272567788 7.893425e-02 0.179396013 -1.92077188 11.95118
#> 41             -0.551538280 2.833572e-02 0.094452391 -2.49322677 11.95118
#> 42             -0.422217281 1.716283e-01 0.261971825 -1.45426074 11.95118
#> 43             -1.427545054 2.486003e-03 0.018421927 -3.81349988 11.95118
#> 44              0.539173151 3.741792e-01 0.467724003  0.92317400 11.95118
#> 45             -0.272402734 2.161159e-01 0.300161028 -1.30604595 11.95118
#> 46              0.624524702 1.056262e-01 0.211252489  1.75054336 11.95118
#> 47              1.728739897 4.768072e-05 0.002121745  6.18612345 11.95118
#> 48              0.539953527 3.389687e-02 0.100553906  2.39509063 11.95118
#> 49              0.219577713 1.199604e-01 0.219501792  1.67459383 11.95118
#> 50              0.891591206 1.107782e-02 0.050353715  3.00163285 11.95118
#>     diff.status     sd.FBS    sem.FBS sd.6h.DMSO sem.6h.DMSO     lfcSE
#> 1          null 0.33770902 0.16885451 0.37455345  0.18727673 0.2175204
#> 2          null 0.17037927 0.08518964 0.18901271  0.09450635 0.1358168
#> 3          null 0.34957014 0.17478507 0.26123347  0.13061673 0.1903941
#> 4  unresponsive 0.07993710 0.03996855 0.11347830  0.05673915 0.1147463
#> 5          null 0.13466320 0.06733160 0.32510829  0.16255415 0.1494019
#> 6          null 0.29666970 0.14833485 0.29204586  0.14602293 0.1747399
#> 7  unresponsive 0.63505961 0.31752980 0.38959597  0.19479799 0.2404600
#> 8          null 1.32765216 0.66382608 0.50741743  0.25370872 0.5457127
#> 9          null 0.50369065 0.25184532 0.00000000  0.00000000 0.2023417
#> 10         null 0.48908126 0.24454063 0.30959834  0.15479917 0.2487822
#> 11 unresponsive 0.67804080 0.33902040 0.64352969  0.32176485 0.4030673
#> 12         null 0.50294265 0.25147132 0.82899627  0.41449813 0.3140250
#> 13         null 0.16778697 0.08389349 0.15423219  0.07711609 0.1210364
#> 14         null 0.21297945 0.10648973 0.34006066  0.17003033 0.1964864
#> 15         null 0.20984911 0.10492456 0.36327101  0.18163550 0.1872122
#> 16          FBS 0.27412743 0.13706371 0.62522412  0.31261206 0.2763906
#> 17         null 0.47967372 0.23983686 0.17929906  0.08964953 0.1994528
#> 18 unresponsive 0.09101198 0.04550599 0.29034987  0.14517494 0.1517601
#> 19         null 0.03040255 0.01520128 0.09066907  0.04533454 0.1024319
#> 20 unresponsive 0.14616929 0.07308464 0.28852850  0.14426425 0.1519629
#> 21 unresponsive 0.12350189 0.06175095 0.23324599  0.11662300 0.1360475
#> 22         null 0.25075066 0.12537533 0.27372493  0.13686247 0.1841545
#> 23         null 0.18772589 0.09386294 0.59276487  0.29638244 0.2768030
#> 24         null 0.53480813 0.26740407 0.19785994  0.09892997 0.2247935
#> 25      6h.DMSO 0.83624527 0.41812263 0.24554775  0.12277387 0.3190594
#> 26         null 0.20110101 0.10055050 0.05618530  0.02809265 0.1337758
#> 27         null 0.11008082 0.05504041 0.45387931  0.22693966 0.2471240
#> 28         null 0.18797445 0.09398722 0.60054988  0.30027494 0.2402053
#> 29         null 0.38289535 0.19144768 0.12953782  0.06476891 0.1692547
#> 30         null 0.25813388 0.12906694 1.02173853  0.51086926 0.3684780
#> 31         null 0.47621831 0.23810916 0.38895134  0.19447567 0.2672547
#> 32 unresponsive 0.23375225 0.11687612 0.31318490  0.15659245 0.1685061
#> 33         null 1.03856869 0.51928434 0.25297686  0.12648843 0.5004606
#> 34         null 0.67968290 0.33984145 0.15492840  0.07746420 0.2766906
#> 35 unresponsive 0.55981879 0.27990939 0.69368990  0.34684495 0.2998215
#> 36         null 0.17682314 0.08841157 0.12672504  0.06336252 0.1222170
#> 37         null 0.19731984 0.09865992 0.46283986  0.23141993 0.1802654
#> 38         null 0.25297972 0.12648986 0.27202837  0.13601419 0.1715847
#> 39 unresponsive 0.25503161 0.12751581 0.25365561  0.12682780 0.1614089
#> 40         null 0.22751850 0.11375925 0.10200115  0.05100057 0.1419053
#> 41         null 0.58362160 0.29181080 0.20032282  0.10016141 0.2212146
#> 42         null 0.71268832 0.35634416 0.35935010  0.17967505 0.2903312
#> 43      6h.DMSO 0.93566401 0.46783201 0.40307134  0.20153567 0.3743399
#> 44         null 1.02299176 0.51149588 1.20518188  0.60259094 0.5840428
#> 45         null 0.28819641 0.14409820 0.38708289  0.19354145 0.2085706
#> 46         null 0.33966999 0.16983499 0.50741743  0.25370872 0.3567605
#> 47          FBS 0.66953381 0.33476691 0.33098965  0.16549483 0.2794545
#> 48         null 0.57578543 0.28789271 0.27687148  0.13843574 0.2254418
#> 49         null 0.13717455 0.06858727 0.17479043  0.08739521 0.1311230
#> 50         null 0.35710360 0.17855180 0.77252649  0.38626324 0.2970354

# Results with the whole protein annotation (when available)
get.results(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
            contrast = 1,
            protein.info.columns = "all")
#>       prot.id basemean.log2 log2.mean.FBS log2.mean.6h.DMSO
#> 1   protein.1      18.54165      18.26954          18.81376
#> 2   protein.2      22.76659      22.43601          23.09716
#> 3   protein.3      16.20940      16.57051          15.84830
#> 4   protein.4      21.19592      21.16429          21.22754
#> 5   protein.5      19.30638      19.18238          19.43038
#> 6   protein.6      15.49009      15.63217          15.34802
#> 7   protein.7      17.00778      17.02135          16.99422
#> 8   protein.8      12.11628      12.69995          11.53260
#> 9   protein.9      11.59300      12.08864          11.09737
#> 10 protein.10      12.84557      13.11600          12.57513
#> 11 protein.11      14.22426      14.25200          14.19653
#> 12 protein.12      14.18435      14.40897          13.95974
#> 13 protein.14      18.17955      18.29240          18.06671
#> 14 protein.15      15.15596      15.29837          15.01355
#> 15 protein.16      16.30529      16.39731          16.21326
#> 16 protein.17      13.42623      14.22905          12.62341
#> 17 protein.18      17.70510      17.48411          17.92609
#> 18 protein.19      15.97148      15.95095          15.99202
#> 19 protein.20      19.39488      19.48232          19.30743
#> 20 protein.21      18.47346      18.47229          18.47463
#> 21 protein.22      16.49923      16.49822          16.50023
#> 22 protein.23      15.42045      15.55763          15.28327
#> 23 protein.24      14.96077      15.07249          14.84904
#> 24 protein.25      18.93194      18.66327          19.20061
#> 25 protein.26      17.13326      16.50530          17.76121
#> 26 protein.27      17.49123      17.68725          17.29522
#> 27 protein.28      11.30874      11.15241          11.46507
#> 28 protein.29      16.36207      16.73584          15.98829
#> 29 protein.30      17.36296      17.05715          17.66877
#> 30 protein.31      12.96287      13.30513          12.62060
#> 31 protein.32      13.33568      13.13449          13.53687
#> 32 protein.33      15.43228      15.42325          15.44131
#> 33 protein.34      12.63760      11.88392          13.39128
#> 34 protein.35      14.31802      14.63020          14.00584
#> 35 protein.36      16.00716      16.04804          15.96629
#> 36 protein.37      19.58236      19.50252          19.66220
#> 37 protein.38      16.53489      16.62487          16.44492
#> 38 protein.39      17.04712      17.27062          16.82361
#> 39 protein.40      16.23806      16.25504          16.22108
#> 40 protein.41      16.32023      16.18395          16.45652
#> 41 protein.42      15.59224      15.31647          15.86801
#> 42 protein.43      16.78640      16.57529          16.99751
#> 43 protein.44      14.07439      13.36062          14.78817
#> 44 protein.45      15.83634      16.10593          15.56675
#> 45 protein.46      15.08760      14.95140          15.22380
#> 46 protein.47      11.84487      12.15713          11.53260
#> 47 protein.48      17.49097      18.35534          16.62660
#> 48 protein.49      17.19600      17.46597          16.92602
#> 49 protein.50      15.79326      15.90305          15.68347
#> 50 protein.51      15.32080      15.76659          14.87500
#>    log2.Fold_FBS.vs.6h.DMSO      p.value        padj   statistic       df
#> 1              -0.544224676 2.788667e-02 0.094452391 -2.50194806 11.95118
#> 2              -0.661148700 3.908730e-04 0.004885913 -4.86794370 11.95118
#> 3               0.722211468 2.579070e-03 0.018421927  3.79324426 11.95118
#> 4              -0.063254105 5.916181e-01 0.704307234 -0.55125200 11.95118
#> 5              -0.247994864 1.229210e-01 0.219501792 -1.65991770 11.95118
#> 6               0.284145196 1.299883e-01 0.224117783  1.62610346 11.95118
#> 7               0.027126811 9.120523e-01 0.954600916  0.11281216 11.95118
#> 8               1.167347745 5.376507e-02 0.128012081  2.13912514 11.95118
#> 9               0.991273547 3.709303e-04 0.004885913  4.89900805 11.95118
#> 10              0.540865019 5.051674e-02 0.126291840  2.17405019 11.95118
#> 11              0.055474587 8.928236e-01 0.954600916  0.13763107 11.95118
#> 12              0.449233938 1.781768e-01 0.262024729  1.43056764 11.95118
#> 13              0.225695196 8.696017e-02 0.183348825  1.86468854 11.95118
#> 14              0.284825720 1.729014e-01 0.261971825  1.44959543 11.95118
#> 15              0.184057322 3.450198e-01 0.442333030  0.98314791 11.95118
#> 16              1.605633047 8.486982e-05 0.002121745  5.80929006 11.95118
#> 17             -0.441985756 4.686072e-02 0.123317671 -2.21599204 11.95118
#> 18             -0.041072472 7.912847e-01 0.899187125 -0.27064076 11.95118
#> 19              0.174882092 1.135883e-01 0.218438960  1.70730119 11.95118
#> 20             -0.002342062 9.879577e-01 0.988477955 -0.01541206 11.95118
#> 21             -0.002006187 9.884780e-01 0.988477955 -0.01474622 11.95118
#> 22              0.274353614 1.621873e-01 0.261592441  1.48980115 11.95118
#> 23              0.223451682 4.352896e-01 0.530840975  0.80725895 11.95118
#> 24             -0.537343331 3.418833e-02 0.100553906 -2.39038682 11.95118
#> 25             -1.255916962 1.991072e-03 0.018421927 -3.93631105 11.95118
#> 26              0.392025553 1.264107e-02 0.052671111  2.93046720 11.95118
#> 27             -0.312661079 2.299120e-01 0.310691916 -1.26519923 11.95118
#> 28              0.747543162 9.025440e-03 0.050141332  3.11210100 11.95118
#> 29             -0.611619251 3.578412e-03 0.022365077 -3.61360374 11.95118
#> 30              0.684529841 8.800744e-02 0.183348825  1.85772225 11.95118
#> 31             -0.402377887 1.581353e-01 0.261592441 -1.50559713 11.95118
#> 32             -0.018062115 9.164169e-01 0.954600916 -0.10718969 11.95118
#> 33             -1.507356091 1.086806e-02 0.050353715 -3.01193758 11.95118
#> 34              0.624360902 4.356655e-02 0.121018205  2.25653070 11.95118
#> 35              0.081753522 7.897580e-01 0.899187125  0.27267397 11.95118
#> 36             -0.159672887 2.159763e-01 0.300161028 -1.30647011 11.95118
#> 37              0.179953720 3.379315e-01 0.442333030  0.99827111 11.95118
#> 38              0.447016255 2.306848e-02 0.088724919  2.60522179 11.95118
#> 39              0.033956247 8.369197e-01 0.929910803  0.21037408 11.95118
#> 40             -0.272567788 7.893425e-02 0.179396013 -1.92077188 11.95118
#> 41             -0.551538280 2.833572e-02 0.094452391 -2.49322677 11.95118
#> 42             -0.422217281 1.716283e-01 0.261971825 -1.45426074 11.95118
#> 43             -1.427545054 2.486003e-03 0.018421927 -3.81349988 11.95118
#> 44              0.539173151 3.741792e-01 0.467724003  0.92317400 11.95118
#> 45             -0.272402734 2.161159e-01 0.300161028 -1.30604595 11.95118
#> 46              0.624524702 1.056262e-01 0.211252489  1.75054336 11.95118
#> 47              1.728739897 4.768072e-05 0.002121745  6.18612345 11.95118
#> 48              0.539953527 3.389687e-02 0.100553906  2.39509063 11.95118
#> 49              0.219577713 1.199604e-01 0.219501792  1.67459383 11.95118
#> 50              0.891591206 1.107782e-02 0.050353715  3.00163285 11.95118
#>     diff.status     sd.FBS    sem.FBS sd.6h.DMSO sem.6h.DMSO     lfcSE
#> 1          null 0.33770902 0.16885451 0.37455345  0.18727673 0.2175204
#> 2          null 0.17037927 0.08518964 0.18901271  0.09450635 0.1358168
#> 3          null 0.34957014 0.17478507 0.26123347  0.13061673 0.1903941
#> 4  unresponsive 0.07993710 0.03996855 0.11347830  0.05673915 0.1147463
#> 5          null 0.13466320 0.06733160 0.32510829  0.16255415 0.1494019
#> 6          null 0.29666970 0.14833485 0.29204586  0.14602293 0.1747399
#> 7  unresponsive 0.63505961 0.31752980 0.38959597  0.19479799 0.2404600
#> 8          null 1.32765216 0.66382608 0.50741743  0.25370872 0.5457127
#> 9          null 0.50369065 0.25184532 0.00000000  0.00000000 0.2023417
#> 10         null 0.48908126 0.24454063 0.30959834  0.15479917 0.2487822
#> 11 unresponsive 0.67804080 0.33902040 0.64352969  0.32176485 0.4030673
#> 12         null 0.50294265 0.25147132 0.82899627  0.41449813 0.3140250
#> 13         null 0.16778697 0.08389349 0.15423219  0.07711609 0.1210364
#> 14         null 0.21297945 0.10648973 0.34006066  0.17003033 0.1964864
#> 15         null 0.20984911 0.10492456 0.36327101  0.18163550 0.1872122
#> 16          FBS 0.27412743 0.13706371 0.62522412  0.31261206 0.2763906
#> 17         null 0.47967372 0.23983686 0.17929906  0.08964953 0.1994528
#> 18 unresponsive 0.09101198 0.04550599 0.29034987  0.14517494 0.1517601
#> 19         null 0.03040255 0.01520128 0.09066907  0.04533454 0.1024319
#> 20 unresponsive 0.14616929 0.07308464 0.28852850  0.14426425 0.1519629
#> 21 unresponsive 0.12350189 0.06175095 0.23324599  0.11662300 0.1360475
#> 22         null 0.25075066 0.12537533 0.27372493  0.13686247 0.1841545
#> 23         null 0.18772589 0.09386294 0.59276487  0.29638244 0.2768030
#> 24         null 0.53480813 0.26740407 0.19785994  0.09892997 0.2247935
#> 25      6h.DMSO 0.83624527 0.41812263 0.24554775  0.12277387 0.3190594
#> 26         null 0.20110101 0.10055050 0.05618530  0.02809265 0.1337758
#> 27         null 0.11008082 0.05504041 0.45387931  0.22693966 0.2471240
#> 28         null 0.18797445 0.09398722 0.60054988  0.30027494 0.2402053
#> 29         null 0.38289535 0.19144768 0.12953782  0.06476891 0.1692547
#> 30         null 0.25813388 0.12906694 1.02173853  0.51086926 0.3684780
#> 31         null 0.47621831 0.23810916 0.38895134  0.19447567 0.2672547
#> 32 unresponsive 0.23375225 0.11687612 0.31318490  0.15659245 0.1685061
#> 33         null 1.03856869 0.51928434 0.25297686  0.12648843 0.5004606
#> 34         null 0.67968290 0.33984145 0.15492840  0.07746420 0.2766906
#> 35 unresponsive 0.55981879 0.27990939 0.69368990  0.34684495 0.2998215
#> 36         null 0.17682314 0.08841157 0.12672504  0.06336252 0.1222170
#> 37         null 0.19731984 0.09865992 0.46283986  0.23141993 0.1802654
#> 38         null 0.25297972 0.12648986 0.27202837  0.13601419 0.1715847
#> 39 unresponsive 0.25503161 0.12751581 0.25365561  0.12682780 0.1614089
#> 40         null 0.22751850 0.11375925 0.10200115  0.05100057 0.1419053
#> 41         null 0.58362160 0.29181080 0.20032282  0.10016141 0.2212146
#> 42         null 0.71268832 0.35634416 0.35935010  0.17967505 0.2903312
#> 43      6h.DMSO 0.93566401 0.46783201 0.40307134  0.20153567 0.3743399
#> 44         null 1.02299176 0.51149588 1.20518188  0.60259094 0.5840428
#> 45         null 0.28819641 0.14409820 0.38708289  0.19354145 0.2085706
#> 46         null 0.33966999 0.16983499 0.50741743  0.25370872 0.3567605
#> 47          FBS 0.66953381 0.33476691 0.33098965  0.16549483 0.2794545
#> 48         null 0.57578543 0.28789271 0.27687148  0.13843574 0.2254418
#> 49         null 0.13717455 0.06858727 0.17479043  0.08739521 0.1311230
#> 50         null 0.35710360 0.17855180 0.77252649  0.38626324 0.2970354