heatmap.contrasts

Plots an heatmap of the log2(FoldChange) for differentially expressed proteins

heatmap.contrasts(
  DEprot.analyses.object,
  contrasts = NULL,
  top.n = NULL,
  distance.method = "euclidean",
  clustering.method = "complete",
  high.color = "firebrick",
  low.color = "#2166AC",
  mid.color = "white",
  na.color = "gray",
  cell.border.color = NA,
  cell.border.width = 0.5,
  title = NULL,
  show.protein.names = FALSE,
  protein.names.pattern = NULL,
  use.uncorrected.pvalue = FALSE
)

Arguments

DEprot.analyses.object

An object of class DEprot.analyses.

contrasts

Numeric vector indicating the position of the contrast to use for the plotting. Default: NULL (all contrasts).

top.n

Numeric value indicated the top differentially expressed proteins to consider for each contrast selected. The rank is based on the product of log2Fc and -log10Padj. Default: NULL (all differential proteins).

distance.method

The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given. Default: "euclidean".

clustering.method

The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). Default: complete"}.}\item{high.color}{String indicating the color to use for up-regulated protein FoldChanges in the plot. Default: \code{"indianred"}.}\item{low.color}{String indicating the color to use for down-regulated protein FoldChanges in the plot. Default: \code{"#2166AC" (blue).

mid.color

String indicating the color to use for the 0 in the plots. Default: "white".

na.color

String indicating the color to use for the NA values in the heatmap. Default: "gray".

cell.border.color

String indicating the color to use for the individual cells border. Default: NA (no border).

cell.border.width

Numeric value indicating the width of the cell border line. Ignored when cell.border.color = NA. Default: 0.5.

title

String indicating the title to use, markdown formatting supported. Default: NULL (automatic title).

show.protein.names

Logical value to indicate whether the protein names should be displayed. Default: FALSE.

protein.names.pattern

Character indicating a regular expression to remove from the protein IDs. Default" NULL, no alterations in the protein IDs.

use.uncorrected.pvalue

Logical value indicating whether it should be used the normal p-value instead of the adjusted one (differential proteins numbers are recomputed). Default: FALSE, padj is used.

Value

A DEprot.contrast.heatmap object, which contains the heatmap (ggplot) and the hclust object used to order the rows.