plot.counts

Plots a volcano plot log2(FoldChange) x -log10(p.adjusted) of differential expression results

# S3 method for class 'counts'
plot(
  DEprot.object,
  which.data = "imputed",
  violin.color = "darkorange",
  max.line.color = "indianred",
  min.line.color = "steelblue",
  title = NULL,
  subtitle = NULL,
  convert.log2 = TRUE
)

Arguments

DEprot.object

An object of class DEprot.analyses.

which.data

String indicating which type of counts should be used. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: "imputed".

violin.color

String indicating the color to use for the violin plots. Default: "darkorange".

max.line.color

String indicating the color to use for up-regulated proteins in the plots. Default: "indianred".

min.line.color

String indicating the color to use for up-regulated proteins in the plots. Default: "steelblue".

title

String indicating the title to use. Default: NULL.

subtitle

String indicating the subtitle to use. Default: NULL.

convert.log2

Logical value to define whether counts should be log2 transformed. Default: TRUE.

Value

A ggplot object.