randomize.missing.values

Function that allows to assign random values (from the bottom-X percent of the LFQ values distribution) to proteins that show missing values for all the replicates of a specific condition.

randomize.missing.values(
  DEprot.object,
  group.column,
  percentage.missing = 100,
  tail.percentage = 3,
  seed = floor(runif(n = 1, min = 0, max = 50000)),
  verbose = TRUE
)

Arguments

DEprot.object

A DEprot object, as generated by load.counts.

group.column

String indicating a column for the metadata table which indicates the IDs of the conditions within which the missing values should be tested.

percentage.missing

Numeric value between 0 and 100 (included) indicated the minimal percentage of missing values per group required to apply the imputation. For instance, if percentage.missing = 75 and the sample group contains 4 samples, imputation will be applied when at least 3 out of 4 values are missing. Default: 100 (all values are missing in the group).

tail.percentage

Numeric value between 0 and 100 (excluded) indicating the threshold percentage of the distribution from which random values should be sampled. Default: 3 (percent).

seed

Numeric value value indicating the random seed used for the imputation. By default a random integer between 0 and 50000 is selected (if verbose = TRUE a message is printed indicating the seed used).

verbose

Logical value to define whether the message indicating which random seed was used should be printed. Default: TRUE.

Value

A DEprot object in which values for proteins which display a number of NAs above a user-defined threshold have been substituted in the normalized counts table with bottom-distribution random values.