remove.undetected.proteins

This function allows for the removal of all the proteins that occur below a certain threshold (detected in at least n samples).

remove.undetected.proteins(
  DEprot.object,
  min.n.samples = 3,
  which.data = "normalized"
)

Arguments

DEprot.object

A DEprot object.

min.n.samples

Numeric value indicating the minimum number of sample in which a protein must be detected to be kept. Default: 3.

which.data

String indicating which type of counts should be used to define the proteins to remove. One among: 'raw', 'normalized', 'norm', 'imputed', 'imp'. Default: "normalized".

Value

A DEprot object.

Author

Sebastian Gregoricchio

Examples

dpo <- load.counts(counts = DEprot::unimputed.counts, metadata = DEprot::sample.config, log.base = 2)
#> The counts matrix contained 295 rows with only NA values.
#> The latter have been removed from the matrix.
dpo <- remove.undetected.proteins(DEprot.object = dpo, min.n.samples = 3, which.data = "raw")
#> These proteins (n=331) will be removed:
#> protein.13, protein.110, protein.118, protein.123, protein.126, protein.136, protein.143, protein.156, protein.195, protein.232, protein.284, protein.327, protein.361, protein.385, protein.486, protein.543, protein.674, protein.807, protein.832, protein.965, protein.1052, protein.1131, protein.1132, protein.1293, protein.1298, protein.1328, protein.1329, protein.1365, protein.1406, protein.1473, protein.1488, protein.1514, protein.1566, protein.1583, protein.1589, protein.1599, protein.1603, protein.1678, protein.1687, protein.1721, protein.1731, protein.1732, protein.1760, protein.1805, protein.1821, protein.1844, protein.1914, protein.1933, protein.1949, protein.1964, protein.1970, protein.1977, protein.2082, protein.2112, protein.2171, protein.2252, protein.2260, protein.2267, protein.2269, protein.2302, protein.2359, protein.2364, protein.2379, protein.2480, protein.2491, protein.2635, protein.2666, protein.2764, protein.2767, protein.2828, protein.2848, protein.2857, protein.2876, protein.2973, protein.2974, protein.2977, protein.3041, protein.3042, protein.3076, protein.3091, protein.3141, protein.3152, protein.3160, protein.3168, protein.3169, protein.3189, protein.3192, protein.3214, protein.3382, protein.3442, protein.3454, protein.3457, protein.3500, protein.3519, protein.3583, protein.3603, protein.3606, protein.3647, protein.3652, protein.3655, protein.3660, protein.3664, protein.3678, protein.3682, protein.3782, protein.3786, protein.3885, protein.3886, protein.3931, protein.3944, protein.3991, protein.4013, protein.4076, protein.4136, protein.4267, protein.4293, protein.4319, protein.4744, protein.4860, protein.4882, protein.4916, protein.4923, protein.4925, protein.4959, protein.4986, protein.5011, protein.5131, protein.5164, protein.5178, protein.5184, protein.5205, protein.5234, protein.5351, protein.5398, protein.5530, protein.5577, protein.5616, protein.5617, protein.5797, protein.5876, protein.5901, protein.5937, protein.5949, protein.6009, protein.6039, protein.6062, protein.6067, protein.6119, protein.6136, protein.6142, protein.6188, protein.6262, protein.6267, protein.6302, protein.6363, protein.6429, protein.6543, protein.6608, protein.6626, protein.6655, protein.6731, protein.6732, protein.6744, protein.6750, protein.6832, protein.6837, protein.6844, protein.6862, protein.6865, protein.6877, protein.6941, protein.6943, protein.6958, protein.7073, protein.7136, protein.7168, protein.7186, protein.7201, protein.7205, protein.7217, protein.7219, protein.7222, protein.7245, protein.7277, protein.7279, protein.7368, protein.7387, protein.7410, protein.7514, protein.7564, protein.7578, protein.7628, protein.7639, protein.7714, protein.7749, protein.7753, protein.7781, protein.7787, protein.7802, protein.7842, protein.7866, protein.7903, protein.7979, protein.7983, protein.8008, protein.8036, protein.8080, protein.8103, protein.8120, protein.8179, protein.8230, protein.8293, protein.8326, protein.8333, protein.8360, protein.8410, protein.8476, protein.8479, protein.8499, protein.8635, protein.8843, protein.8884, protein.8964, protein.8998, protein.9001, protein.9036, protein.9042, protein.9143, protein.9164, protein.9199, protein.9232, protein.9233, protein.9265, protein.9275, protein.9338, protein.9368, protein.9462, protein.9491, protein.9632, protein.9648, protein.9658, protein.9697, protein.9725, protein.9746, protein.9773, protein.9834, protein.9875, protein.9887, protein.9929, protein.9936, protein.10053, protein.10089, protein.10108, protein.10124, protein.10167, protein.10200, protein.10237, protein.10285, protein.10309, protein.10350, protein.10391, protein.10501, protein.10565, protein.10634, protein.10652, protein.10654, protein.10670, protein.10682, protein.10749, protein.10759, protein.10794, protein.10809, protein.10869, protein.10897, protein.10940, protein.10950, protein.10976, protein.11010, protein.11053, protein.11056, protein.11059, protein.11077, protein.11103, protein.11211, protein.11214, protein.11281, protein.11290, protein.11307, protein.11340, protein.11360, protein.11407, protein.11431, protein.11576, protein.11591, protein.11634, protein.11656, protein.11682, protein.11685, protein.11829, protein.11883, protein.11900, protein.11935, protein.11970, protein.11976, protein.12022, protein.12134, protein.12204, protein.12254, protein.12266, protein.12385, protein.12398, protein.12422, protein.12424, protein.12495, protein.12515, protein.12527, protein.12544, protein.12564, protein.12566, protein.12730, protein.12811, protein.12830, protein.12994, protein.13066, protein.13068, protein.13073, protein.13075, protein.13076, protein.13104, protein.13215, protein.13221