Test toolbox

Set of objects and datasets used for the package tests.

data("test.toolbox")

Format

A list of objects:

test.unimputed.lfq

table of a subset of 50 proteins from the unimputed.counts table. Usable in combination with sample.config

dpo.raw

Object of class DEprot with unimputed data loaded as raw

dpo.norm

Object of class DEprot containing MBQN normalized data

dpo.imp

Object of class DEprot containing missForest imputed data

diff.exp.limma

Object of class DEprot.analyses containing limma defined differential expression for two contrasts

geneset

Dummy geneset TERM2GENE table for GSEA and ORA analyses containing 3 sets of proteins

gsea.results

Object of class DEprot.enrichResult result of GSEA analyses on contrast 1

ora.results

Object of class DEprot.enrichResult result of ORA analyses on contrast 1 and differential category '6h.DMSO'

Source

Simulated data