timecourse.enrichment

Performs an OverRepresentation Analysis (ORA) independently for each cluster of a time-course analysis, using as background (universe) all the proteins that were tested. It returns the combined results and a dotplot in which the size of each dot reflects the enrichment, the color its significance, and the number written inside the dot the count of proteins found in that geneset.

timecourse.enrichment(
  DEprot.timecourse.object,
  TERM2GENE,
  clusters = NULL,
  top.n = NULL,
  universe = NULL,
  gsub.pattern.prot.id = NULL,
  min.cluster.size = 5,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  dotplot.n = 5,
  size.by = "FoldEnrichment",
  size.range = c(5, 14),
  number.size = 2.7,
  show.numbers = TRUE,
  max.term.length = 60,
  viridis.option = "rocket"
)

Arguments

DEprot.timecourse.object

An object of class DEprot.timecourse.

TERM2GENE

Data.frame containing two columns 'gs_name' (IDs of the gene sets) and 'gene_symbol' (indicating the gene IDs). No default.

clusters

Numeric value (or vector) indicating the cluster(s) to analyze. Default: NULL (all the clusters).

top.n

Numeric value indicating how many best-ranked proteins of each cluster should be used for the enrichment. Default: NULL (all the proteins of the cluster).

universe

Character vector indicating the background gene list. Default: NULL, meaning that all the proteins tested by analyze.timecourse are used. This is almost always the correct background for a proteomics experiment, since only the quantified proteins could have been detected as trending.

gsub.pattern.prot.id

String indicating a pattern to be passed to gsub and to remove from the prot.id. Default: NULL (no changes in the IDs).

min.cluster.size

Numeric value indicating the minimal number of proteins required to run the enrichment on a cluster. Default: 5.

pvalueCutoff

Numeric value indicating the adjusted pvalue cutoff on enrichment tests to report. Default: 0.05.

qvalueCutoff

Numeric value indicating the qvalue cutoff on enrichment tests to report as significant. Default: 0.05.

pAdjustMethod

String indicating the method to use for the p-value adjustment. One among "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: "BH".

dotplot.n

Numeric value indicating the maximum number of genesets displayed per cluster in the dotplot. Default: 5.

size.by

String indicating the metric mapped on the size of the dots. One among: 'FoldEnrichment' or 'GeneRatio'. Default: "FoldEnrichment".

size.range

Numeric vector of two elements indicating the minimal and maximal size of the dots. They must stay large enough for the numbers to fit inside. Default: c(5, 14).

number.size

Numeric value indicating the font size of the count displayed inside the dots. Default: 2.7.

show.numbers

Logic value indicating whether the number of proteins should be written inside the dots. Default: TRUE.

max.term.length

Numeric value indicating the maximal number of characters of the geneset names displayed on the y-axis (longer names are truncated). Default: 60.

viridis.option

String indicating the viridis palette used for the significance color scale. Default: "rocket".

Value

An object of class DEprot.timecourse.enrichment.

See also

Author

Sebastian Gregoricchio