get.results

Simplifies the access to the differential expression results table

get.results(DEprot.analyses.object, contrast = 1)

Arguments

DEprot.analyses.object

An object of class DEprot.analyses.

contrast

Number indicating the position of the contrast to use for the plotting.

Value

A data frame.

Author

Sebastian Gregoricchio

Examples

get.results(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
            contrast = 1)
#>       prot.id basemean.log2 log2.mean.FBS log2.mean.6h.DMSO
#> 1   protein.1      18.54165      18.26954          18.81376
#> 2   protein.2      22.76659      22.43601          23.09716
#> 3   protein.3      16.20940      16.57051          15.84830
#> 4   protein.4      21.19592      21.16429          21.22754
#> 5   protein.5      19.30638      19.18238          19.43038
#> 6   protein.6      15.49009      15.63217          15.34802
#> 7   protein.7      17.00778      17.02135          16.99422
#> 8   protein.8      12.01085      12.67978          11.34192
#> 9   protein.9      11.70320      12.11284          11.29356
#> 10 protein.10      12.84557      13.11600          12.57513
#> 11 protein.11      14.22429      14.25200          14.19658
#> 12 protein.12      14.18435      14.40897          13.95974
#> 13 protein.14      18.17955      18.29240          18.06671
#> 14 protein.15      15.15596      15.29837          15.01355
#> 15 protein.16      16.30529      16.39731          16.21326
#> 16 protein.17      13.42704      14.22905          12.62503
#> 17 protein.18      17.70510      17.48411          17.92609
#> 18 protein.19      15.97148      15.95095          15.99202
#> 19 protein.20      19.39488      19.48232          19.30743
#> 20 protein.21      18.47346      18.47229          18.47463
#> 21 protein.22      16.49923      16.49822          16.50023
#> 22 protein.23      15.42045      15.55763          15.28327
#> 23 protein.24      14.96077      15.07249          14.84904
#> 24 protein.25      18.93194      18.66327          19.20061
#> 25 protein.26      17.13337      16.50552          17.76121
#> 26 protein.27      17.49123      17.68725          17.29522
#> 27 protein.28      11.30868      11.15237          11.46499
#> 28 protein.29      16.36207      16.73584          15.98829
#> 29 protein.30      17.36296      17.05715          17.66877
#> 30 protein.31      12.95905      13.29750          12.62060
#> 31 protein.32      13.33568      13.13449          13.53687
#> 32 protein.33      15.43228      15.42325          15.44131
#> 33 protein.34      12.51808      11.64488          13.39128
#> 34 protein.35      14.31868      14.63020          14.00716
#> 35 protein.36      16.00716      16.04804          15.96629
#> 36 protein.37      19.58236      19.50252          19.66220
#> 37 protein.38      16.53489      16.62487          16.44492
#> 38 protein.39      17.04712      17.27062          16.82361
#> 39 protein.40      16.23806      16.25504          16.22108
#> 40 protein.41      16.32023      16.18395          16.45652
#> 41 protein.42      15.59224      15.31647          15.86801
#> 42 protein.43      16.78640      16.57529          16.99751
#> 43 protein.44      14.07332      13.35847          14.78817
#> 44 protein.45      15.83421      16.10506          15.56337
#> 45 protein.46      15.08760      14.95140          15.22380
#> 46 protein.47      11.73205      12.12218          11.34192
#> 47 protein.48      17.49097      18.35534          16.62660
#> 48 protein.49      17.19600      17.46597          16.92602
#> 49 protein.50      15.79326      15.90305          15.68347
#> 50 protein.51      15.31780      15.76659          14.86900
#>    log2.Fold_FBS.vs.6h.DMSO      p.value        padj   statistic       df
#> 1              -0.544224676 4.446987e-02 0.145470742 -2.24263355 12.06676
#> 2              -0.661148700 7.974475e-04 0.013290791 -4.43910773 12.06676
#> 3               0.722211468 6.405979e-03 0.040037367  3.29077590 12.06676
#> 4              -0.063254105 6.372689e-01 0.758653397 -0.48368394 12.06676
#> 5              -0.247994864 1.790676e-01 0.313015024 -1.42655041 12.06676
#> 6               0.284145196 1.906222e-01 0.317703590  1.38668565 12.06676
#> 7               0.027126811 9.257991e-01 0.964374097  0.09509653 12.06676
#> 8               1.337855591 3.954715e-02 0.141239818  2.30746781 12.06676
#> 9               0.819279937 7.988055e-03 0.044378083  3.17209156 12.06676
#> 10              0.540865019 7.965538e-02 0.181034948  1.91382791 12.06676
#> 11              0.055415822 9.108760e-01 0.964374097  0.11430493 12.06676
#> 12              0.449233938 2.508475e-01 0.358353637  1.20622827 12.06676
#> 13              0.225695196 1.326549e-01 0.265309778  1.61263439 12.06676
#> 14              0.284825720 2.058405e-01 0.321625856  1.33713172 12.06676
#> 15              0.184057322 3.710648e-01 0.488243106  0.92907186 12.06676
#> 16              1.604011896 1.672282e-04 0.004180705  5.36130818 12.06676
#> 17             -0.441985756 6.793353e-02 0.161746495 -2.00492739 12.06676
#> 18             -0.041072472 8.102526e-01 0.934673774 -0.24542392 12.06676
#> 19              0.174882092 1.680704e-01 0.313015024  1.46654862 12.06676
#> 20             -0.002342062 9.897984e-01 0.989995252 -0.01305350 12.06676
#> 21             -0.002006187 9.899953e-01 0.989995252 -0.01280167 12.06676
#> 22              0.274353614 2.054858e-01 0.321625856  1.33825199 12.06676
#> 23              0.223451682 4.847482e-01 0.591156320  0.72082560 12.06676
#> 24             -0.537343331 5.706394e-02 0.145470742 -2.10349816 12.06676
#> 25             -1.255690444 2.551616e-03 0.031895205 -3.79006706 12.06676
#> 26              0.392025553 2.048325e-02 0.085346871  2.66587998 12.06676
#> 27             -0.312622839 2.504860e-01 0.358353637 -1.20720217 12.06676
#> 28              0.747543162 1.822979e-02 0.082862674  2.72875417 12.06676
#> 29             -0.611619251 5.338804e-03 0.040037367 -3.38897597 12.06676
#> 30              0.676896166 1.305593e-01 0.265309778  1.62229794 12.06676
#> 31             -0.402377887 1.712110e-01 0.313015024 -1.45490637 12.06676
#> 32             -0.018062115 9.247476e-01 0.964374097 -0.09644873 12.06676
#> 33             -1.746402464 5.733484e-03 0.040037367 -3.35051481 12.06676
#> 34              0.623047502 5.320937e-02 0.145470742  2.14272250 12.06676
#> 35              0.081753522 8.225129e-01 0.934673774  0.22925178 12.06676
#> 36             -0.159672887 2.948743e-01 0.398478721 -1.09504166 12.06676
#> 37              0.179953720 4.248215e-01 0.544642946  0.82602449 12.06676
#> 38              0.447016255 3.078859e-02 0.118417641  2.44474238 12.06676
#> 39              0.033956247 8.595854e-01 0.955094918  0.18072221 12.06676
#> 40             -0.272567788 1.266737e-01 0.265309778 -1.64058136 12.06676
#> 41             -0.551538280 5.818830e-02 0.145470742 -2.09252453 12.06676
#> 42             -0.422217281 2.146619e-01 0.325245258 -1.30976982 12.06676
#> 43             -1.429699926 3.472477e-03 0.034724771 -3.62189811 12.06676
#> 44              0.541692837 4.842914e-01 0.591156320  0.72159558 12.06676
#> 45             -0.272402734 2.675244e-01 0.371561679 -1.16245919 12.06676
#> 46              0.780256459 5.592580e-02 0.145470742  2.11481463 12.06676
#> 47              1.728739897 8.506102e-05 0.004180705  5.78360156 12.06676
#> 48              0.539953527 5.268929e-02 0.145470742  2.14821808 12.06676
#> 49              0.219577713 1.815487e-01 0.313015024  1.41781296 12.06676
#> 50              0.897597139 1.548880e-02 0.077443989  2.81647283 12.06676
#>     diff.status     sd.FBS    sem.FBS sd.6h.DMSO sem.6h.DMSO     lfcSE
#> 1          null 0.33770902 0.16885451 0.37455345  0.18727673 0.2426721
#> 2          null 0.17037927 0.08518964 0.18901271  0.09450635 0.1489373
#> 3          null 0.34957014 0.17478507 0.26123347  0.13061673 0.2194654
#> 4  unresponsive 0.07993710 0.03996855 0.11347830  0.05673915 0.1307757
#> 5          null 0.13466320 0.06733160 0.32510829  0.16255415 0.1738423
#> 6          null 0.29666970 0.14833485 0.29204586  0.14602293 0.2049096
#> 7  unresponsive 0.63505961 0.31752980 0.38959597  0.19479799 0.2852555
#> 8          null 1.33930886 0.66965443 0.48911443  0.24455722 0.5797938
#> 9          null 0.46495273 0.23247637 0.39238086  0.19619043 0.2582775
#> 10         null 0.48908126 0.24454063 0.30959834  0.15479917 0.2826090
#> 11 unresponsive 0.67804080 0.33902040 0.64351653  0.32175826 0.4848069
#> 12         null 0.50294265 0.25147132 0.82899627  0.41449813 0.3724286
#> 13         null 0.16778697 0.08389349 0.15423219  0.07711609 0.1399543
#> 14         null 0.21297945 0.10648973 0.34006066  0.17003033 0.2130125
#> 15         null 0.20984911 0.10492456 0.36327101  0.18163550 0.1981088
#> 16          FBS 0.27412743 0.13706371 0.62627213  0.31313607 0.2991829
#> 17         null 0.47967372 0.23983686 0.17929906  0.08964953 0.2204498
#> 18 unresponsive 0.09101198 0.04550599 0.29034987  0.14517494 0.1673532
#> 19         null 0.03040255 0.01520128 0.09066907  0.04533454 0.1192474
#> 20 unresponsive 0.14616929 0.07308464 0.28852850  0.14426425 0.1794202
#> 21 unresponsive 0.12350189 0.06175095 0.23324599  0.11662300 0.1567129
#> 22         null 0.25075066 0.12537533 0.27372493  0.13686247 0.2050089
#> 23         null 0.18772589 0.09386294 0.59276487  0.29638244 0.3099941
#> 24         null 0.53480813 0.26740407 0.19785994  0.09892997 0.2554522
#> 25      6h.DMSO 0.83648570 0.41824285 0.24554775  0.12277387 0.3313109
#> 26         null 0.20110101 0.10055050 0.05618530  0.02809265 0.1470530
#> 27         null 0.11000434 0.05500217 0.45391247  0.22695623 0.2589648
#> 28         null 0.18797445 0.09398722 0.60054988  0.30027494 0.2739504
#> 29         null 0.38289535 0.19144768 0.12953782  0.06476891 0.1804732
#> 30         null 0.26312288 0.13156144 1.02173853  0.51086926 0.4172453
#> 31         null 0.47621831 0.23810916 0.38895134  0.19447567 0.2765662
#> 32 unresponsive 0.23375225 0.11687612 0.31318490  0.15659245 0.1872717
#> 33      6h.DMSO 1.09501867 0.54750934 0.25297686  0.12648843 0.5212341
#> 34         null 0.67968290 0.33984145 0.15592387  0.07796194 0.2907738
#> 35 unresponsive 0.55981879 0.27990939 0.69368990  0.34684495 0.3566102
#> 36         null 0.17682314 0.08841157 0.12672504  0.06336252 0.1458144
#> 37         null 0.19731984 0.09865992 0.46283986  0.23141993 0.2178552
#> 38         null 0.25297972 0.12648986 0.27202837  0.13601419 0.1828480
#> 39 unresponsive 0.25503161 0.12751581 0.25365561  0.12682780 0.1878919
#> 40         null 0.22751850 0.11375925 0.10200115  0.05100057 0.1661410
#> 41         null 0.58362160 0.29181080 0.20032282  0.10016141 0.2635755
#> 42         null 0.71268832 0.35634416 0.35935010  0.17967505 0.3223599
#> 43      6h.DMSO 0.93362384 0.46681192 0.40307134  0.20153567 0.3947378
#> 44         null 1.02332435 0.51166217 1.20407622  0.60203811 0.7506876
#> 45         null 0.28819641 0.14409820 0.38708289  0.19354145 0.2343332
#> 46         null 0.35679778 0.17839889 0.48911443  0.24455722 0.3689479
#> 47          FBS 0.66953381 0.33476691 0.33098965  0.16549483 0.2989037
#> 48         null 0.57578543 0.28789271 0.27687148  0.13843574 0.2513495
#> 49         null 0.13717455 0.06858727 0.17479043  0.08739521 0.1548707
#> 50         null 0.35710360 0.17855180 0.76922716  0.38461358 0.3186955