get.results

Simplifies the access to the differential expression results table

get.results(DEprot.analyses.object, contrast = 1)

Arguments

DEprot.analyses.object

An object of class DEprot.analyses.

contrast

Number indicating the position of the contrast to use for the plotting.

Value

A data frame.

Author

Sebastian Gregoricchio

Examples

get.results(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
            contrast = 1)
#>       prot.id basemean.log2 log2.mean.FBS log2.mean.6h.DMSO
#> 1   protein.1      18.54165      18.26954          18.81376
#> 2   protein.2      22.76659      22.43601          23.09716
#> 3   protein.3      16.20940      16.57051          15.84830
#> 4   protein.4      21.19592      21.16429          21.22754
#> 5   protein.5      19.30638      19.18238          19.43038
#> 6   protein.6      15.49009      15.63217          15.34802
#> 7   protein.7      17.00778      17.02135          16.99422
#> 8   protein.8      11.87376      12.65016          11.09737
#> 9   protein.9      11.70222      12.11088          11.29356
#> 10 protein.10      12.84557      13.11600          12.57513
#> 11 protein.11      14.22425      14.25200          14.19650
#> 12 protein.12      14.18435      14.40897          13.95974
#> 13 protein.14      18.17955      18.29240          18.06671
#> 14 protein.15      15.15596      15.29837          15.01355
#> 15 protein.16      16.30529      16.39731          16.21326
#> 16 protein.17      13.42185      14.22905          12.61466
#> 17 protein.18      17.70510      17.48411          17.92609
#> 18 protein.19      15.97148      15.95095          15.99202
#> 19 protein.20      19.39488      19.48232          19.30743
#> 20 protein.21      18.47346      18.47229          18.47463
#> 21 protein.22      16.49923      16.49822          16.50023
#> 22 protein.23      15.42045      15.55763          15.28327
#> 23 protein.24      14.96077      15.07249          14.84904
#> 24 protein.25      18.93194      18.66327          19.20061
#> 25 protein.26      17.13375      16.50629          17.76121
#> 26 protein.27      17.49123      17.68725          17.29522
#> 27 protein.28      11.27083      11.12714          11.41453
#> 28 protein.29      16.36207      16.73584          15.98829
#> 29 protein.30      17.36296      17.05715          17.66877
#> 30 protein.31      12.96078      13.30097          12.62060
#> 31 protein.32      13.33568      13.13449          13.53687
#> 32 protein.33      15.43228      15.42325          15.44131
#> 33 protein.34      12.51808      11.64488          13.39128
#> 34 protein.35      14.31787      14.63020          14.00554
#> 35 protein.36      16.00716      16.04804          15.96629
#> 36 protein.37      19.58236      19.50252          19.66220
#> 37 protein.38      16.53489      16.62487          16.44492
#> 38 protein.39      17.04712      17.27062          16.82361
#> 39 protein.40      16.23806      16.25504          16.22108
#> 40 protein.41      16.32023      16.18395          16.45652
#> 41 protein.42      15.59224      15.31647          15.86801
#> 42 protein.43      16.78640      16.57529          16.99751
#> 43 protein.44      14.07344      13.35872          14.78817
#> 44 protein.45      15.83507      16.10344          15.56671
#> 45 protein.46      15.08760      14.95140          15.22380
#> 46 protein.47      11.88140      12.18183          11.58097
#> 47 protein.48      17.49097      18.35534          16.62660
#> 48 protein.49      17.19600      17.46597          16.92602
#> 49 protein.50      15.79326      15.90305          15.68347
#> 50 protein.51      15.32203      15.76659          14.87747
#>    log2.Fold_FBS.vs.6h.DMSO      p.value        padj   statistic       df
#> 1              -0.544224676 4.419724e-02 0.147324135 -2.24579120 12.07841
#> 2              -0.661148700 7.775611e-04 0.012959352 -4.45218092 12.07841
#> 3               0.722211468 6.335490e-03 0.039596814  3.29608857 12.07841
#> 4              -0.063254105 6.360341e-01 0.757183428 -0.48546386 12.07841
#> 5              -0.247994864 1.781241e-01 0.300683944 -1.42981469 12.07841
#> 6               0.284145196 1.898650e-01 0.306233791  1.38915567 12.07841
#> 7               0.027126811 9.257124e-01 0.964283790  0.09520609 12.07841
#> 8               1.552794212 1.664295e-02 0.075437294  2.77738183 12.07841
#> 9               0.817318033 8.050242e-03 0.044723566  3.16734165 12.07841
#> 10              0.540865019 7.946343e-02 0.189198631  1.91504810 12.07841
#> 11              0.055498459 9.106780e-01 0.964283790  0.11455771 12.07841
#> 12              0.449233938 2.504338e-01 0.357762610  1.20728397 12.07841
#> 13              0.225695196 1.314823e-01 0.262964555  1.61791192 12.07841
#> 14              0.284825720 2.051047e-01 0.310764649  1.33938365 12.07841
#> 15              0.184057322 3.701827e-01 0.487082462  0.93081120 12.07841
#> 16              1.614383944 1.461073e-04 0.003652681  5.44231137 12.07841
#> 17             -0.441985756 6.752780e-02 0.168819507 -2.00814391 12.07841
#> 18             -0.041072472 8.097973e-01 0.934489365 -0.24602064 12.07841
#> 19              0.174882092 1.663532e-01 0.300683944  1.47290137 12.07841
#> 20             -0.002342062 9.897760e-01 0.989967982 -0.01308193 12.07841
#> 21             -0.002006187 9.899680e-01 0.989967982 -0.01283631 12.07841
#> 22              0.274353614 2.047099e-01 0.310764649  1.34063405 12.07841
#> 23              0.223451682 4.842864e-01 0.595298866  0.72158205 12.07841
#> 24             -0.537343331 5.676163e-02 0.152350278 -2.10626977 12.07841
#> 25             -1.254929083 2.554394e-03 0.031929927 -3.78856832 12.07841
#> 26              0.392025553 2.016685e-02 0.077564803  2.67390096 12.07841
#> 27             -0.287390167 1.777717e-01 0.300683944 -1.43106940 12.07841
#> 28              0.747543162 1.810495e-02 0.075437294  2.73205199 12.07841
#> 29             -0.611619251 5.260189e-03 0.039596814 -3.39628824 12.07841
#> 30              0.680364057 1.283265e-01 0.262964555  1.63262916 12.07841
#> 31             -0.402377887 1.707148e-01 0.300683944 -1.45664478 12.07841
#> 32             -0.018062115 9.245932e-01 0.964283790 -0.09664529 12.07841
#> 33             -1.746402464 5.702092e-03 0.039596814 -3.35280317 12.07841
#> 34              0.624668434 5.241155e-02 0.152350278  2.15094912 12.07841
#> 35              0.081753522 8.223506e-01 0.934489365  0.22946059 12.07841
#> 36             -0.159672887 2.934500e-01 0.396554110 -1.09838375 12.07841
#> 37              0.179953720 4.240715e-01 0.543681471  0.82737193 12.07841
#> 38              0.447016255 3.048102e-02 0.108860798  2.44991137 12.07841
#> 39              0.033956247 8.593012e-01 0.954779101  0.18108866 12.07841
#> 40             -0.272567788 1.258052e-01 0.262964555 -1.64461723 12.07841
#> 41             -0.551538280 5.789311e-02 0.152350278 -2.09517551 12.07841
#> 42             -0.422217281 2.142068e-01 0.315010025 -1.31108840 12.07841
#> 43             -1.429449978 3.456210e-03 0.034562101 -3.62364099 12.07841
#> 44              0.536731316 4.881451e-01 0.595298866  0.71509151 12.07841
#> 45             -0.272402734 2.668340e-01 0.370602782 -1.16417399 12.07841
#> 46              0.600861462 1.273686e-01 0.262964555  1.63715918 12.07841
#> 47              1.728739897 8.390209e-05 0.003652681  5.78991895 12.07841
#> 48              0.539953527 5.239668e-02 0.152350278  2.15110772 12.07841
#> 49              0.219577713 1.804104e-01 0.300683944  1.42172514 12.07841
#> 50              0.889119988 1.657708e-02 0.075437294  2.77951567 12.07841
#>     diff.status     sd.FBS    sem.FBS sd.6h.DMSO sem.6h.DMSO     lfcSE
#> 1          null 0.33770902 0.16885451 0.37455345  0.18727673 0.2423309
#> 2          null 0.17037927 0.08518964 0.18901271  0.09450635 0.1485000
#> 3          null 0.34957014 0.17478507 0.26123347  0.13061673 0.2191117
#> 4  unresponsive 0.07993710 0.03996855 0.11347830  0.05673915 0.1302962
#> 5          null 0.13466320 0.06733160 0.32510829  0.16255415 0.1734455
#> 6          null 0.29666970 0.14833485 0.29204586  0.14602293 0.2045453
#> 7  unresponsive 0.63505961 0.31752980 0.38959597  0.19479799 0.2849273
#> 8          null 1.35841606 0.67920803 0.00000000  0.00000000 0.5590856
#> 9          null 0.46802668 0.23401334 0.39238086  0.19619043 0.2580454
#> 10         null 0.48908126 0.24454063 0.30959834  0.15479917 0.2824289
#> 11 unresponsive 0.67804080 0.33902040 0.64353505  0.32176752 0.4844585
#> 12         null 0.50294265 0.25147132 0.82899627  0.41449813 0.3721030
#> 13         null 0.16778697 0.08389349 0.15423219  0.07711609 0.1394978
#> 14         null 0.21297945 0.10648973 0.34006066  0.17003033 0.2126543
#> 15         null 0.20984911 0.10492456 0.36327101  0.18163550 0.1977386
#> 16          FBS 0.27412743 0.13706371 0.61982932  0.30991466 0.2966357
#> 17         null 0.47967372 0.23983686 0.17929906  0.08964953 0.2200967
#> 18 unresponsive 0.09101198 0.04550599 0.29034987  0.14517494 0.1669473
#> 19         null 0.03040255 0.01520128 0.09066907  0.04533454 0.1187331
#> 20 unresponsive 0.14616929 0.07308464 0.28852850  0.14426425 0.1790304
#> 21 unresponsive 0.12350189 0.06175095 0.23324599  0.11662300 0.1562900
#> 22         null 0.25075066 0.12537533 0.27372493  0.13686247 0.2046447
#> 23         null 0.18772589 0.09386294 0.59276487  0.29638244 0.3096691
#> 24         null 0.53480813 0.26740407 0.19785994  0.09892997 0.2551161
#> 25      6h.DMSO 0.83729511 0.41864755 0.24554775  0.12277387 0.3312410
#> 26         null 0.20110101 0.10055050 0.05618530  0.02809265 0.1466118
#> 27         null 0.05953900 0.02976950 0.47885024  0.23942512 0.2008220
#> 28         null 0.18797445 0.09398722 0.60054988  0.30027494 0.2736197
#> 29         null 0.38289535 0.19144768 0.12953782  0.06476891 0.1800846
#> 30         null 0.26075748 0.13037874 1.02173853  0.51086926 0.4167291
#> 31         null 0.47621831 0.23810916 0.38895134  0.19447567 0.2762361
#> 32 unresponsive 0.23375225 0.11687612 0.31318490  0.15659245 0.1868908
#> 33      6h.DMSO 1.09501867 0.54750934 0.25297686  0.12648843 0.5208783
#> 34         null 0.67968290 0.33984145 0.15545499  0.07772749 0.2904153
#> 35 unresponsive 0.55981879 0.27990939 0.69368990  0.34684495 0.3562857
#> 36         null 0.17682314 0.08841157 0.12672504  0.06336252 0.1453708
#> 37         null 0.19731984 0.09865992 0.46283986  0.23141993 0.2175004
#> 38         null 0.25297972 0.12648986 0.27202837  0.13601419 0.1824622
#> 39 unresponsive 0.25503161 0.12751581 0.25365561  0.12682780 0.1875117
#> 40         null 0.22751850 0.11375925 0.10200115  0.05100057 0.1657333
#> 41         null 0.58362160 0.29181080 0.20032282  0.10016141 0.2632420
#> 42         null 0.71268832 0.35634416 0.35935010  0.17967505 0.3220357
#> 43      6h.DMSO 0.93385969 0.46692985 0.40307134  0.20153567 0.3944789
#> 44         null 1.02395432 0.51197716 1.20516649  0.60258325 0.7505771
#> 45         null 0.28819641 0.14409820 0.38708289  0.19354145 0.2339880
#> 46         null 0.33001255 0.16500628 0.55848983  0.27924492 0.3670147
#> 47          FBS 0.66953381 0.33476691 0.33098965  0.16549483 0.2985776
#> 48         null 0.57578543 0.28789271 0.27687148  0.13843574 0.2510118
#> 49         null 0.13717455 0.06858727 0.17479043  0.08739521 0.1544446
#> 50         null 0.35710360 0.17855180 0.77393409  0.38696705 0.3198831