get.results

Simplifies the access to the differential expression results table

get.results(DEprot.analyses.object, contrast = 1)

Arguments

DEprot.analyses.object

An object of class DEprot.analyses.

contrast

Number indicating the position of the contrast to use for the plotting.

Value

A data frame.

Author

Sebastian Gregoricchio

Examples

get.results(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
            contrast = 1)
#>       prot.id basemean.log2 log2.mean.FBS log2.mean.6h.DMSO
#> 1   protein.1      18.54165      18.26954          18.81376
#> 2   protein.2      22.76659      22.43601          23.09716
#> 3   protein.3      16.20940      16.57051          15.84830
#> 4   protein.4      21.19592      21.16429          21.22754
#> 5   protein.5      19.30638      19.18238          19.43038
#> 6   protein.6      15.49009      15.63217          15.34802
#> 7   protein.7      17.00778      17.02135          16.99422
#> 8   protein.8      11.87376      12.65016          11.09737
#> 9   protein.9      11.86314      12.14531          11.58097
#> 10 protein.10      12.84557      13.11600          12.57513
#> 11 protein.11      14.22417      14.25200          14.19634
#> 12 protein.12      14.18435      14.40897          13.95974
#> 13 protein.14      18.17955      18.29240          18.06671
#> 14 protein.15      15.15596      15.29837          15.01355
#> 15 protein.16      16.30529      16.39731          16.21326
#> 16 protein.17      13.44074      14.22905          12.65244
#> 17 protein.18      17.70510      17.48411          17.92609
#> 18 protein.19      15.97148      15.95095          15.99202
#> 19 protein.20      19.39488      19.48232          19.30743
#> 20 protein.21      18.47346      18.47229          18.47463
#> 21 protein.22      16.49923      16.49822          16.50023
#> 22 protein.23      15.42045      15.55763          15.28327
#> 23 protein.24      14.96077      15.07249          14.84904
#> 24 protein.25      18.93194      18.66327          19.20061
#> 25 protein.26      17.13383      16.50644          17.76121
#> 26 protein.27      17.49123      17.68725          17.29522
#> 27 protein.28      11.27083      11.12714          11.41453
#> 28 protein.29      16.36207      16.73584          15.98829
#> 29 protein.30      17.36296      17.05715          17.66877
#> 30 protein.31      12.96098      13.30136          12.62060
#> 31 protein.32      13.33568      13.13449          13.53687
#> 32 protein.33      15.43228      15.42325          15.44131
#> 33 protein.34      12.51808      11.64488          13.39128
#> 34 protein.35      14.31653      14.63020          14.00286
#> 35 protein.36      16.00716      16.04804          15.96629
#> 36 protein.37      19.58236      19.50252          19.66220
#> 37 protein.38      16.53489      16.62487          16.44492
#> 38 protein.39      17.04712      17.27062          16.82361
#> 39 protein.40      16.23806      16.25504          16.22108
#> 40 protein.41      16.32023      16.18395          16.45652
#> 41 protein.42      15.59224      15.31647          15.86801
#> 42 protein.43      16.78640      16.57529          16.99751
#> 43 protein.44      14.07177      13.35538          14.78817
#> 44 protein.45      15.83161      16.10207          15.56115
#> 45 protein.46      15.08760      14.95140          15.22380
#> 46 protein.47      11.57676      12.05615          11.09737
#> 47 protein.48      17.49097      18.35534          16.62660
#> 48 protein.49      17.19600      17.46597          16.92602
#> 49 protein.50      15.79326      15.90305          15.68347
#> 50 protein.51      15.31934      15.76659          14.87209
#>    log2.Fold_FBS.vs.6h.DMSO      p.value        padj   statistic       df
#> 1              -0.544224676 4.427896e-02 0.147596524 -2.24481885 12.07603
#> 2              -0.661148700 7.839601e-04 0.013066002 -4.44777890 12.07603
#> 3               0.722211468 6.356766e-03 0.039729791  3.29441547 12.07603
#> 4              -0.063254105 6.364553e-01 0.757684869 -0.48485487 12.07603
#> 5              -0.247994864 1.784311e-01 0.301313074 -1.42874053 12.07603
#> 6               0.284145196 1.901038e-01 0.306619057  1.38836673 12.07603
#> 7               0.027126811 9.257379e-01 0.964310362  0.09517368 12.07603
#> 8               1.552794212 1.665947e-02 0.075587042  2.77693409 12.07603
#> 9               0.564337613 7.913609e-02 0.180861891  1.91745883 12.07603
#> 10              0.540865019 7.957923e-02 0.180861891  1.91424404 12.07603
#> 11              0.055655604 9.104436e-01 0.964310362  0.11486028 12.07603
#> 12              0.449233938 2.505459e-01 0.357922740  1.20699387 12.07603
#> 13              0.225695196 1.318760e-01 0.263751963  1.61612059 12.07603
#> 14              0.284825720 2.053350e-01 0.311113673  1.33867000 12.07603
#> 15              0.184057322 3.704641e-01 0.487452764  0.93025197 12.07603
#> 16              1.576610596 2.202141e-04 0.005505352  5.19219755 12.07603
#> 17             -0.441985756 6.765273e-02 0.169131816 -2.00713191 12.07603
#> 18             -0.041072472 8.099475e-01 0.934540359 -0.24582308 12.07603
#> 19              0.174882092 1.669414e-01 0.301313074  1.47070658 12.07603
#> 20             -0.002342062 9.897833e-01 0.989977074 -0.01307262 12.07603
#> 21             -0.002006187 9.899771e-01 0.989977074 -0.01282472 12.07603
#> 22              0.274353614 2.049547e-01 0.311113673  1.33987328 12.07603
#> 23              0.223451682 4.844189e-01 0.591137863  0.72136320 12.07603
#> 24             -0.537343331 5.685128e-02 0.152578809 -2.10542667 12.07603
#> 25             -1.254770376 2.565241e-03 0.032065510 -3.78643199 12.07603
#> 26              0.392025553 2.027113e-02 0.077965881  2.67119555 12.07603
#> 27             -0.287390167 1.780118e-01 0.301313074 -1.43023152 12.07603
#> 28              0.747543162 1.814089e-02 0.075587042  2.73106617 12.07603
#> 29             -0.611619251 5.284939e-03 0.039729791 -3.39389710 12.07603
#> 30              0.680761357 1.281858e-01 0.263751963  1.63331631 12.07603
#> 31             -0.402377887 1.708603e-01 0.301313074 -1.45612638 12.07603
#> 32             -0.018062115 9.246432e-01 0.964310362 -0.09658143 12.07603
#> 33             -1.746402464 5.709602e-03 0.039729791 -3.35223060 12.07603
#> 34              0.627342413 5.160509e-02 0.152578809  2.15965994 12.07603
#> 35              0.081753522 8.223955e-01 0.934540359  0.22940250 12.07603
#> 36             -0.159672887 2.939277e-01 0.397199581 -1.09725524 12.07603
#> 37              0.179953720 4.243063e-01 0.543982455  0.82694692 12.07603
#> 38              0.447016255 3.057894e-02 0.109210499  2.44822492 12.07603
#> 39              0.033956247 8.593931e-01 0.954881259  0.18096967 12.07603
#> 40             -0.272567788 1.260893e-01 0.263751963 -1.64327954 12.07603
#> 41             -0.551538280 5.797995e-02 0.152578809 -2.09437530 12.07603
#> 42             -0.422217281 2.143351e-01 0.315198665 -1.31071100 12.07603
#> 43             -1.432789054 3.356138e-03 0.033561382 -3.63978326 12.07603
#> 44              0.540921400 4.847330e-01 0.591137863  0.72083365 12.07603
#> 45             -0.272402734 2.670459e-01 0.370897116 -1.16364168 12.07603
#> 46              0.958783227 1.704892e-02 0.075587042  2.76449747 12.07603
#> 47              1.728739897 8.421089e-05 0.004210544  5.78807316 12.07603
#> 48              0.539953527 5.248375e-02 0.152578809  2.15022528 12.07603
#> 49              0.219577713 1.807878e-01 0.301313074  1.42041489 12.07603
#> 50              0.894501892 1.588347e-02 0.075587042  2.80259432 12.07603
#>     diff.status
#> 1          null
#> 2          null
#> 3          null
#> 4  unresponsive
#> 5          null
#> 6          null
#> 7  unresponsive
#> 8          null
#> 9          null
#> 10         null
#> 11 unresponsive
#> 12         null
#> 13         null
#> 14         null
#> 15         null
#> 16          FBS
#> 17         null
#> 18 unresponsive
#> 19         null
#> 20 unresponsive
#> 21 unresponsive
#> 22         null
#> 23         null
#> 24         null
#> 25      6h.DMSO
#> 26         null
#> 27         null
#> 28         null
#> 29         null
#> 30         null
#> 31         null
#> 32 unresponsive
#> 33      6h.DMSO
#> 34         null
#> 35 unresponsive
#> 36         null
#> 37         null
#> 38         null
#> 39 unresponsive
#> 40         null
#> 41         null
#> 42         null
#> 43      6h.DMSO
#> 44         null
#> 45         null
#> 46         null
#> 47          FBS
#> 48         null
#> 49         null
#> 50         null