get.results

Simplifies the access to the differential expression results table

get.results(DEprot.analyses.object, contrast = 1)

Arguments

DEprot.analyses.object

An object of class DEprot.analyses.

contrast

Number indicating the position of the contrast to use for the plotting.

Value

A data frame.

Author

Sebastian Gregoricchio

Examples

get.results(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
            contrast = 1)
#>       prot.id basemean.log2 log2.mean.FBS log2.mean.6h.DMSO
#> 1   protein.1      18.54165      18.26954          18.81376
#> 2   protein.2      22.76659      22.43601          23.09716
#> 3   protein.3      16.20940      16.57051          15.84830
#> 4   protein.4      21.19592      21.16429          21.22754
#> 5   protein.5      19.30638      19.18238          19.43038
#> 6   protein.6      15.49009      15.63217          15.34802
#> 7   protein.7      17.00778      17.02135          16.99422
#> 8   protein.8      11.87376      12.65016          11.09737
#> 9   protein.9      11.72727      12.11812          11.33641
#> 10 protein.10      12.84557      13.11600          12.57513
#> 11 protein.11      14.22484      14.25200          14.19768
#> 12 protein.12      14.18435      14.40897          13.95974
#> 13 protein.14      18.17955      18.29240          18.06671
#> 14 protein.15      15.15596      15.29837          15.01355
#> 15 protein.16      16.30529      16.39731          16.21326
#> 16 protein.17      13.42609      14.22905          12.62313
#> 17 protein.18      17.70510      17.48411          17.92609
#> 18 protein.19      15.97148      15.95095          15.99202
#> 19 protein.20      19.39488      19.48232          19.30743
#> 20 protein.21      18.47346      18.47229          18.47463
#> 21 protein.22      16.49923      16.49822          16.50023
#> 22 protein.23      15.42045      15.55763          15.28327
#> 23 protein.24      14.96077      15.07249          14.84904
#> 24 protein.25      18.93194      18.66327          19.20061
#> 25 protein.26      17.13336      16.50550          17.76121
#> 26 protein.27      17.49123      17.68725          17.29522
#> 27 protein.28      11.30874      11.15241          11.46507
#> 28 protein.29      16.36207      16.73584          15.98829
#> 29 protein.30      17.36296      17.05715          17.66877
#> 30 protein.31      12.96276      13.30493          12.62060
#> 31 protein.32      13.33568      13.13449          13.53687
#> 32 protein.33      15.43228      15.42325          15.44131
#> 33 protein.34      12.61617      11.84107          13.39128
#> 34 protein.35      14.31455      14.63020          13.99890
#> 35 protein.36      16.00716      16.04804          15.96629
#> 36 protein.37      19.58236      19.50252          19.66220
#> 37 protein.38      16.53489      16.62487          16.44492
#> 38 protein.39      17.04712      17.27062          16.82361
#> 39 protein.40      16.23806      16.25504          16.22108
#> 40 protein.41      16.32023      16.18395          16.45652
#> 41 protein.42      15.59224      15.31647          15.86801
#> 42 protein.43      16.78640      16.57529          16.99751
#> 43 protein.44      14.07384      13.35951          14.78817
#> 44 protein.45      15.83512      16.10640          15.56384
#> 45 protein.46      15.08760      14.95140          15.22380
#> 46 protein.47      11.87796      12.18045          11.57546
#> 47 protein.48      17.49097      18.35534          16.62660
#> 48 protein.49      17.19600      17.46597          16.92602
#> 49 protein.50      15.79326      15.90305          15.68347
#> 50 protein.51      15.32116      15.76659          14.87573
#>    log2.Fold_FBS.vs.6h.DMSO      p.value        padj   statistic       df
#> 1              -0.544224676 0.0444484004 0.148161335 -2.24239600 12.09042
#> 2              -0.661148700 0.0008055397 0.013425662 -4.43069516 12.09042
#> 3               0.722211468 0.0064041072 0.045743623  3.28963325 12.09042
#> 4              -0.063254105 0.6381676162 0.759723353 -0.48235654 12.09042
#> 5              -0.247994864 0.1794757792 0.314106132 -1.42493448 12.09042
#> 6               0.284145196 0.1907978127 0.317996354  1.38593402 12.09042
#> 7               0.027126811 0.9257817883 0.964356029  0.09511495 12.09042
#> 8               1.552794212 0.0166193213 0.075542369  2.77771167 12.09042
#> 9               0.781709560 0.0148522710 0.075542369  2.83813376 12.09042
#> 10              0.540865019 0.0796091963 0.189545705  1.91381882 12.09042
#> 11              0.054322614 0.9125529513 0.964356029  0.11213970 12.09042
#> 12              0.449233938 0.2505710902 0.363932269  1.20685369 12.09042
#> 13              0.225695196 0.1334079897 0.266815979  1.60896700 12.09042
#> 14              0.284825720 0.2059750870 0.321836073  1.33655851 12.09042
#> 15              0.184057322 0.3713295914 0.488591568  0.92846976 12.09042
#> 16              1.605918218 0.0001560723 0.003901808  5.39953220 12.09042
#> 17             -0.441985756 0.0679680224 0.169920056 -2.00425488 12.09042
#> 18             -0.041072472 0.8104859098 0.934571006 -0.24510506 12.09042
#> 19              0.174882092 0.1693977076 0.314106132  1.46142421 12.09042
#> 20             -0.002342062 0.9898082746 0.990011035 -0.01304040 12.09042
#> 21             -0.002006187 0.9900110346 0.990011035 -0.01278095 12.09042
#> 22              0.274353614 0.2056676290 0.321836073  1.33752858 12.09042
#> 23              0.223451682 0.4845874392 0.590960292  0.72105201 12.09042
#> 24             -0.537343331 0.0570201534 0.152977418 -2.10349785 12.09042
#> 25             -1.255717713 0.0025352458 0.031690573 -3.79175772 12.09042
#> 26              0.392025553 0.0206538123 0.079437740  2.66063105 12.09042
#> 27             -0.312661079 0.2547525882 0.363932269 -1.19566295 12.09042
#> 28              0.747543162 0.0181903010 0.075792921  2.72910232 12.09042
#> 29             -0.611619251 0.0053579290 0.044649408 -3.38565204 12.09042
#> 30              0.684324609 0.1262026064 0.264756285  1.64259282 12.09042
#> 31             -0.402377887 0.1711052304 0.314106132 -1.45511531 12.09042
#> 32             -0.018062115 0.9248076118 0.964356029 -0.09636758 12.09042
#> 33             -1.550212035 0.0096263318 0.060164574 -3.07081188 12.09042
#> 34              0.631303906 0.0504316308 0.152977418  2.17221506 12.09042
#> 35              0.081753522 0.8224224857 0.934571006  0.22936106 12.09042
#> 36             -0.159672887 0.2957650232 0.399682464 -1.09283076 12.09042
#> 37              0.179953720 0.4249561074 0.544815522  0.82572144 12.09042
#> 38              0.447016255 0.0308818514 0.110292326  2.44246791 12.09042
#> 39              0.033956247 0.8596947377 0.955216375  0.18057231 12.09042
#> 40             -0.272567788 0.1270830166 0.264756285 -1.63839534 12.09042
#> 41             -0.551538280 0.0581314187 0.152977418 -2.09264852 12.09042
#> 42             -0.422217281 0.2144588531 0.324937656 -1.31024677 12.09042
#> 43             -1.428653638 0.0034827984 0.034827984 -3.61864443 12.09042
#> 44              0.542556619 0.4831797591 0.590960292  0.72342667 12.09042
#> 45             -0.272402734 0.2675652074 0.371618344 -1.16224478 12.09042
#> 46              0.604994596 0.1242588010 0.264756285  1.65195198 12.09042
#> 47              1.728739897 0.0000842069 0.003901808  5.78512736 12.09042
#> 48              0.539953527 0.0526530291 0.152977418  2.14814858 12.09042
#> 49              0.219577713 0.1821815566 0.314106132  1.41543119 12.09042
#> 50              0.890860100 0.0163736781 0.075542369  2.78571995 12.09042
#>     diff.status
#> 1          null
#> 2          null
#> 3          null
#> 4  unresponsive
#> 5          null
#> 6          null
#> 7  unresponsive
#> 8          null
#> 9          null
#> 10         null
#> 11 unresponsive
#> 12         null
#> 13         null
#> 14         null
#> 15         null
#> 16          FBS
#> 17         null
#> 18 unresponsive
#> 19         null
#> 20 unresponsive
#> 21 unresponsive
#> 22         null
#> 23         null
#> 24         null
#> 25      6h.DMSO
#> 26         null
#> 27         null
#> 28         null
#> 29         null
#> 30         null
#> 31         null
#> 32 unresponsive
#> 33         null
#> 34         null
#> 35 unresponsive
#> 36         null
#> 37         null
#> 38         null
#> 39 unresponsive
#> 40         null
#> 41         null
#> 42         null
#> 43      6h.DMSO
#> 44         null
#> 45         null
#> 46         null
#> 47          FBS
#> 48         null
#> 49         null
#> 50         null