get.results

Simplifies the access to the differential expression results table

get.results(DEprot.analyses.object, contrast = 1)

Arguments

DEprot.analyses.object

An object of class DEprot.analyses.

contrast

Number indicating the position of the contrast to use for the plotting.

Value

A data frame.

Author

Sebastian Gregoricchio

Examples

get.results(DEprot.analyses.object = DEprot::test.toolbox$diff.exp.limma,
            contrast = 1)
#>       prot.id basemean.log2 log2.mean.FBS log2.mean.6h.DMSO
#> 1   protein.1      18.54165      18.26954          18.81376
#> 2   protein.2      22.76659      22.43601          23.09716
#> 3   protein.3      16.20940      16.57051          15.84830
#> 4   protein.4      21.19592      21.16429          21.22754
#> 5   protein.5      19.30638      19.18238          19.43038
#> 6   protein.6      15.49009      15.63217          15.34802
#> 7   protein.7      17.00778      17.02135          16.99422
#> 8   protein.8      12.22225      12.77121          11.67328
#> 9   protein.9      11.79314      12.15845          11.42784
#> 10 protein.10      12.84557      13.11600          12.57513
#> 11 protein.11      14.21366      14.25200          14.17533
#> 12 protein.12      14.18435      14.40897          13.95974
#> 13 protein.14      18.17955      18.29240          18.06671
#> 14 protein.15      15.15596      15.29837          15.01355
#> 15 protein.16      16.30529      16.39731          16.21326
#> 16 protein.17      13.42816      14.22905          12.62727
#> 17 protein.18      17.70510      17.48411          17.92609
#> 18 protein.19      15.97148      15.95095          15.99202
#> 19 protein.20      19.39488      19.48232          19.30743
#> 20 protein.21      18.47346      18.47229          18.47463
#> 21 protein.22      16.49923      16.49822          16.50023
#> 22 protein.23      15.42045      15.55763          15.28327
#> 23 protein.24      14.96077      15.07249          14.84904
#> 24 protein.25      18.93194      18.66327          19.20061
#> 25 protein.26      17.11203      16.46284          17.76121
#> 26 protein.27      17.49123      17.68725          17.29522
#> 27 protein.28      11.30990      11.14631          11.47349
#> 28 protein.29      16.36207      16.73584          15.98829
#> 29 protein.30      17.36296      17.05715          17.66877
#> 30 protein.31      12.95736      13.29412          12.62060
#> 31 protein.32      13.33568      13.13449          13.53687
#> 32 protein.33      15.43228      15.42325          15.44131
#> 33 protein.34      13.36773      13.34418          13.39128
#> 34 protein.35      14.28607      14.63020          13.94194
#> 35 protein.36      16.00716      16.04804          15.96629
#> 36 protein.37      19.58236      19.50252          19.66220
#> 37 protein.38      16.53489      16.62487          16.44492
#> 38 protein.39      17.04712      17.27062          16.82361
#> 39 protein.40      16.23806      16.25504          16.22108
#> 40 protein.41      16.32023      16.18395          16.45652
#> 41 protein.42      15.59224      15.31647          15.86801
#> 42 protein.43      16.78640      16.57529          16.99751
#> 43 protein.44      14.04758      13.30699          14.78817
#> 44 protein.45      15.86849      16.18885          15.54814
#> 45 protein.46      15.08760      14.95140          15.22380
#> 46 protein.47      12.49232      12.40296          12.58168
#> 47 protein.48      17.49097      18.35534          16.62660
#> 48 protein.49      17.19600      17.46597          16.92602
#> 49 protein.50      15.78207      15.89672          15.66742
#> 50 protein.51      15.32077      15.76659          14.87495
#>    log2.Fold_FBS.vs.6h.DMSO      p.value        padj  diff.status
#> 1              -0.544224676 4.313645e-02 0.154058733         null
#> 2              -0.661148700 7.684945e-04 0.012808242         null
#> 3               0.722211468 6.089996e-03 0.038062473         null
#> 4              -0.063254105 6.383918e-01 0.778526584 unresponsive
#> 5              -0.247994864 1.779000e-01 0.317678633         null
#> 6               0.284145196 1.885915e-01 0.325157737         null
#> 7               0.027126811 9.252942e-01 0.963848144 unresponsive
#> 8               1.097929751 7.058515e-02 0.185750402         null
#> 9               0.730609437 4.476198e-03 0.036679827         null
#> 10              0.540865019 7.747635e-02 0.193690868         null
#> 11              0.076674081 8.761545e-01 0.963848144 unresponsive
#> 12              0.449233938 2.470489e-01 0.374316461         null
#> 13              0.225695196 1.329680e-01 0.277016600         null
#> 14              0.284825720 2.036061e-01 0.328397012         null
#> 15              0.184057322 3.691331e-01 0.512684801         null
#> 16              1.601773496 1.178507e-04 0.002946268          FBS
#> 17             -0.441985756 6.642049e-02 0.184501368         null
#> 18             -0.041072472 8.099564e-01 0.954821814 unresponsive
#> 19              0.174882092 1.701493e-01 0.315091349         null
#> 20             -0.002342062 9.897719e-01 0.989990662 unresponsive
#> 21             -0.002006187 9.899907e-01 0.989990662 unresponsive
#> 22              0.274353614 2.034179e-01 0.328397012         null
#> 23              0.223451682 4.815671e-01 0.617393749         null
#> 24             -0.537343331 5.544103e-02 0.166184226         null
#> 25             -1.298375673 1.380030e-03 0.017250378      6h.DMSO
#> 26              0.392025553 2.032552e-02 0.092388714         null
#> 27             -0.327178247 1.294868e-01 0.277016600         null
#> 28              0.747543162 1.741187e-02 0.087059335         null
#> 29             -0.611619251 5.135176e-03 0.036679827         null
#> 30              0.673519251 1.292471e-01 0.277016600         null
#> 31             -0.402377887 1.682905e-01 0.315091349         null
#> 32             -0.018062115 9.245091e-01 0.963848144 unresponsive
#> 33             -0.047096006 8.946976e-01 0.963848144 unresponsive
#> 34              0.688263451 3.448549e-02 0.132636510         null
#> 35              0.081753522 8.211468e-01 0.954821814 unresponsive
#> 36             -0.159672887 2.951002e-01 0.421571714         null
#> 37              0.179953720 4.225235e-01 0.555952031         null
#> 38              0.447016255 3.012137e-02 0.125505722         null
#> 39              0.033956247 8.591371e-01 0.963848144 unresponsive
#> 40             -0.272567788 1.258389e-01 0.277016600         null
#> 41             -0.551538280 5.650264e-02 0.166184226         null
#> 42             -0.422217281 2.112159e-01 0.330024914         null
#> 43             -1.481179691 1.990916e-03 0.019909160      6h.DMSO
#> 44              0.640712512 4.032313e-01 0.544907215         null
#> 45             -0.272402734 2.648102e-01 0.389426763         null
#> 46             -0.178725560 5.348115e-01 0.668514433         null
#> 47              1.728739897 7.441234e-05 0.002946268          FBS
#> 48              0.539953527 5.116272e-02 0.166184226         null
#> 49              0.229293759 1.609343e-01 0.315091349         null
#> 50              0.891645555 1.533297e-02 0.085183180         null